[GEN] fix a few bugs identified by test ROM (#669)
* MOVEM register-to-memory does a spurious word read to the final address
* Z80 cannot read from 68000 working RAM
* full 17-bit VRAM address is checked when updating the sprite table cache on VRAM writes
* fix transparent sprite pixels being able to trigger the sprite collision flag (though this test still fails due to timing)

+++ backend/genesis-core/src/api/debug.rs
@@ -80,7 +80,7 @@ pub struct SpriteAttributeEntry {
 
 #[derive(Debug, Clone, Copy)]
 pub struct CopySpriteAttributesResult {
-    pub sprite_table_len: u16,
+    pub sprite_table_len: u32,
     pub top_left_x: u16,
     pub top_left_y: u16,
 }
